backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypdf
anthropic
pydantic
slowapi
python-dotenv
python-jobspy
